IGNITE-25560 Sql. Partition awareness cover extension
review.

+++ modules/sql-engine/src/main/java/org/apache/ignite/internal/sql/engine/util/Cloner.java
@@ -78,7 +78,7 @@ public class Cloner {
 
         private final RelOptCluster cluster;
 
-        private int sourceIndex;
+        private long sourceIndex;
 
         private final Long2ObjectMap<IgniteRel> sources = new Long2ObjectOpenHashMap<>();
